8.2.5 (#303)
- Fix #304 - teambuffs link in charactercard now to go the correct page.
- Refactor src code structure, resolves #297.
- Add Swap weapon button to equip tab. fixes #262.
- Fix artifact set selection dropdown being off screen on 1st load.
