+++ package.json
@@ -1,7 +1,7 @@
 {
   "homepage": "https://frzyc.github.io/genshin-optimizer/",
   "name": "genshin-optimizer",
-  "version": "8.2.4",
+  "version": "8.2.5",
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.7.1",
