+++ src/PageCharacter/Tabs/TabBuild/Components/AssumeFullLevelToggle.tsx
@@ -1,5 +1,5 @@
 import { MenuItem } from "@mui/material"
-import DropdownButton from "../../Components/DropdownMenu/DropdownButton"
+import DropdownButton from "../../../../Components/DropdownMenu/DropdownButton"
 
 const levels = {
   0: <span>No Lv. assumption</span>,
