+++ src/PageCharacter/Tabs/TabBuild/Components/BonusStatsCard.tsx
@@ -1,10 +1,10 @@
 import { CardContent, Divider, ListItem, Typography } from '@mui/material';
 import { useContext } from 'react';
-import CardLight from '../../Components/Card/CardLight';
-import { FieldDisplayList, NodeFieldDisplay } from '../../Components/FieldDisplay';
-import { DataContext } from '../../DataContext';
-import { uiInput as input } from '../../Formula';
-import { NumNode } from '../../Formula/type';
+import CardLight from '../../../../Components/Card/CardLight';
+import { FieldDisplayList, NodeFieldDisplay } from '../../../../Components/FieldDisplay';
+import { DataContext } from '../../../../DataContext';
+import { uiInput as input } from '../../../../Formula';
+import { NumNode } from '../../../../Formula/type';
 export default function BonusStatsCard() {
   const { data, character } = useContext(DataContext)
   const bonusStatsKeys = Object.keys(character?.bonusStats)
