+++ src/PageCharacter/Tabs/TabBuild/common.ts
@@ -1,9 +1,9 @@
-import { forEachNodes, mapFormulas } from "../Formula/internal";
-import { allOperations, constantFold } from "../Formula/optimization";
-import { ConstantNode, NumNode } from "../Formula/type";
-import { constant, customRead, max, min } from "../Formula/utils";
-import { allSlotKeys, ArtifactSetKey } from "../Types/consts";
-import { assertUnreachable, objectKeyMap, objectMap } from "../Util/Util";
+import { forEachNodes, mapFormulas } from "../../../Formula/internal";
+import { allOperations, constantFold } from "../../../Formula/optimization";
+import { ConstantNode, NumNode } from "../../../Formula/type";
+import { constant, customRead, max, min } from "../../../Formula/utils";
+import { allSlotKeys, ArtifactSetKey } from "../../../Types/consts";
+import { assertUnreachable, objectKeyMap, objectMap } from "../../../Util/Util";
 import type { ArtifactBuildData, ArtifactsBySlot, Build, DynStat, PlotData, RequestFilter } from "./background";
 
 type DynMinMax = { [key in string]: MinMax }
