+++ src/PageWeapon/WeaponDisplayCard.tsx
@@ -1,7 +1,6 @@
-import { Lock, LockOpen, SwapHoriz } from "@mui/icons-material"
+import { Lock, LockOpen } from "@mui/icons-material"
 import { Box, Button, ButtonGroup, CardContent, CardHeader, Divider, Grid, ListItem, MenuItem, Typography } from "@mui/material"
 import { useCallback, useContext, useEffect, useMemo, useState } from "react"
-import Assets from "../Assets/Assets"
 import CardDark from "../Components/Card/CardDark"
 import CardLight from "../Components/Card/CardLight"
 import CharacterDropdownButton from "../Components/Character/CharacterDropdownButton"
@@ -10,7 +9,6 @@ import CustomNumberInput, { CustomNumberInputButtonGroupWrapper } from "../Compo
 import DocumentDisplay from "../Components/DocumentDisplay"
 import DropdownButton from "../Components/DropdownMenu/DropdownButton"
 import { FieldDisplayList, NodeFieldDisplay } from "../Components/FieldDisplay"
-import ImgIcon from "../Components/Image/ImgIcon"
 import ModalWrapper from "../Components/ModalWrapper"
 import { Stars } from "../Components/StarDisplay"
 import WeaponSelectionModal from "../Components/Weapon/WeaponSelectionModal"
@@ -26,7 +24,7 @@ import useWeapon from "../ReactHooks/useWeapon"
 import { CharacterKey } from "../Types/consts"
 import { ICachedWeapon } from "../Types/weapon"
 import { clamp } from "../Util/Util"
-import WeaponCard from "./WeaponCard"
+import WeaponSwapBtn from "./WeaponSwapBtn"
 
 type WeaponStatsEditorCardProps = {
   weaponId: string
@@ -44,7 +42,6 @@ export default function WeaponDisplayCard({
   const weapon = useWeapon(propWeaponId)
   const { key = "", level = 0, refinement = 0, ascension = 0, lock, location = "", id } = weapon ?? {}
   const weaponSheet = usePromise(WeaponSheet.get(key), [key])
-  const weaponTypeKey = weaponSheet?.weaponType
 
   const weaponDispatch = useCallback((newWeapon: Partial<ICachedWeapon>) => {
     database.updateWeapon(newWeapon, propWeaponId)
@@ -113,7 +110,7 @@ export default function WeaponDisplayCard({
                   </MenuItem>)}
               </DropdownButton>}
             </ButtonGroup>
-            {data?.get(input.charKey)?.value && <SwapBtn weaponTypeKey={weaponTypeKey} onChangeId={id => database.setWeaponLocation(id, data.get(input.charKey).value as CharacterKey)} />}
+            {data?.get(input.charKey)?.value && <WeaponSwapBtn weaponTypeKey={weaponSheet.weaponType} onChangeId={id => database.setWeaponLocation(id, data.get(input.charKey).value as CharacterKey)} />}
           </Box>
           <Box display="flex" gap={1} flexWrap="wrap" justifyContent="space-between">
             <ButtonGroup sx={{ bgcolor: t => t.palette.contentLight.main }} >
@@ -168,55 +165,4 @@ export default function WeaponDisplayCard({
       </Grid>
     </CardContent>}
   </CardLight ></ModalWrapper>
-}
-function SwapBtn({ onChangeId, weaponTypeKey }) {
-  const { database } = useContext(DatabaseContext)
-  const [show, setShow] = useState(false)
-  const open = () => setShow(true)
-  const close = () => setShow(false)
-
-  const clickHandler = (id) => {
-    onChangeId(id)
-    close()
-  }
-
-  const weaponSheets = usePromise(WeaponSheet.getAll, [])
-
-  const weaponIdList = database.weapons.keys.filter(wKey => {
-    const dbWeapon = database._getWeapon(wKey)
-    if (!dbWeapon) return false
-    if (weaponTypeKey && weaponTypeKey !== weaponSheets?.[dbWeapon.key]?.weaponType) return false
-    return true
-  })
-
-
-  return <>
-    <Button color="info" onClick={open} startIcon={<SwapHoriz />} >SWAP WEAPON</Button>
-    <ModalWrapper open={show} onClose={close} >
-      <CardDark>
-        <CardContent sx={{ py: 1 }}>
-          <Grid container>
-            <Grid item flexGrow={1}>
-              <Typography variant="h6">{weaponTypeKey ? <ImgIcon src={Assets.weaponTypes[weaponTypeKey]} /> : null} Swap Weapon</Typography>
-            </Grid>
-            <Grid item>
-              <CloseButton onClick={close} />
-            </Grid>
-          </Grid>
-        </CardContent>
-        <Divider />
-        <CardContent>
-          <Grid container spacing={1}>
-            {weaponIdList.map(weaponId =>
-              <Grid item key={weaponId} xs={6} sm={6} md={4} lg={3} >
-                <WeaponCard
-                  weaponId={weaponId}
-                  onClick={clickHandler}
-                />
-              </Grid>)}
-          </Grid>
-        </CardContent>
-      </CardDark>
-    </ModalWrapper>
-  </>
 }
