--- src/Util/FormulaTextUtil.tsx
@@ -1,6 +1,0 @@
-/**
- * @deprecated
- */
-export function basicDMGFormulaText(percent: number, stats: any, skillKey: string, elemental?: any) {
-  return <></>
-}
